added score manager (accidently removed sprite models)
