+++ Assets/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b111057284efbd6499ea2f1f42d057bf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
