+++ Assets/Script/destroyobject.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class destroyobject : MonoBehaviour
+{
+    void OnMouseDown()
+    {
+        Destroy(gameObject);
+
+        scoremanager.instance.AddPoint();
+    }
+}
