+++ Assets/Script/destroyobject.cs.meta
