+++ Assets/Script/scoremanager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5aea61e5da62bd243bc66fd65d260749
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
