+++ Assets/trash_game.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 776250d981498354580f7b9c564f2c4f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
