+++ Logs/ApiUpdaterCheck.txt
@@ -1,0 +1,10 @@
+[api-updater (non-obsolete-error-filter)] 2/18/2022 4:09:16 PM : Starting C:/Users/Antonio/Desktop/2020.3.19f1/Editor/Data/Tools/ScriptUpdater/APIUpdater.NonObsoleteApiUpdaterDetector.exe
+[api-updater (non-obsolete-error-filter)] 
+----------------------------------
+jit/startup time      : 664.388ms
+moved types parse time: 56ms
+candidates parse time : 1ms
+C# parse time         : 639ms
+candidates check time : 56ms
+console write time    : 0ms
+
