+++ Logs/Packages-Update.log
@@ -50,3 +50,22 @@ The following packages were updated:
   com.unity.2d.pixel-perfect from version 4.0.1 to 5.0.0
   com.unity.2d.psdimporter from version 4.0.1 to 5.0.1
   com.unity.2d.spriteshape from version 5.0.1 to 6.0.0
+
+=== Fri Feb 18 17:02:32 2022
+
+Packages were changed.
+Update Mode: updateDependencies
+
+The following packages were updated:
+  com.unity.2d.animation from version 6.0.1 to 7.0.3
+  com.unity.2d.pixel-perfect from version 5.0.0 to 5.0.1
+  com.unity.2d.psdimporter from version 5.0.1 to 6.0.2
+  com.unity.2d.spriteshape from version 6.0.0 to 7.0.3
+  com.unity.collab-proxy from version 1.3.9 to 1.15.7
+  com.unity.ide.rider from version 3.0.6 to 3.0.9
+  com.unity.ide.visualstudio from version 2.0.7 to 2.0.14
+  com.unity.ide.vscode from version 1.2.3 to 1.2.4
+  com.unity.test-framework from version 1.1.24 to 1.1.29
+  com.unity.textmeshpro from version 3.0.4 to 3.0.6
+  com.unity.timeline from version 1.5.4 to 1.6.3
+  com.unity.visualscripting from version 1.5.2 to 1.7.6
