+++ Logs/shadercompiler-AssetImportWorker0.log
@@ -1,3 +1,6 @@
-Base path: 'C:/Program Files/Unity 2021.1.6f1/Editor/Data', plugins path 'C:/Program Files/Unity 2021.1.6f1/Editor/Data/PlaybackEngines'
+Base path: 'C:/Users/Antonio/Desktop/2021.2.11f1/Editor/Data', plugins path 'C:/Users/Antonio/Desktop/2021.2.11f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
+Unhandled exception: Protocol error - failed to read magic number (error -2147483644, transferred 0/4)
+
+Quitting shader compiler process
