+++ Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,3 +1,4 @@
-Base path: 'C:/Program Files/Unity 2021.1.6f1/Editor/Data', plugins path 'C:/Program Files/Unity 2021.1.6f1/Editor/Data/PlaybackEngines'
+Base path: 'C:/Users/Antonio/Desktop/2021.2.11f1/Editor/Data', plugins path 'C:/Users/Antonio/Desktop/2021.2.11f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
+Cmd: shutdown
