+++ Packages/manifest.json
@@ -1,20 +1,20 @@
 {
   "dependencies": {
-    "com.unity.2d.animation": "6.0.1",
-    "com.unity.2d.pixel-perfect": "5.0.0",
-    "com.unity.2d.psdimporter": "5.0.1",
+    "com.unity.2d.animation": "7.0.3",
+    "com.unity.2d.pixel-perfect": "5.0.1",
+    "com.unity.2d.psdimporter": "6.0.2",
     "com.unity.2d.sprite": "1.0.0",
-    "com.unity.2d.spriteshape": "6.0.0",
+    "com.unity.2d.spriteshape": "7.0.3",
     "com.unity.2d.tilemap": "1.0.0",
-    "com.unity.collab-proxy": "1.3.9",
-    "com.unity.ide.rider": "3.0.6",
-    "com.unity.ide.visualstudio": "2.0.7",
-    "com.unity.ide.vscode": "1.2.3",
-    "com.unity.test-framework": "1.1.24",
-    "com.unity.textmeshpro": "3.0.4",
-    "com.unity.timeline": "1.5.4",
+    "com.unity.collab-proxy": "1.15.7",
+    "com.unity.ide.rider": "3.0.9",
+    "com.unity.ide.visualstudio": "2.0.14",
+    "com.unity.ide.vscode": "1.2.4",
+    "com.unity.test-framework": "1.1.29",
+    "com.unity.textmeshpro": "3.0.6",
+    "com.unity.timeline": "1.6.3",
     "com.unity.ugui": "1.0.0",
-    "com.unity.visualscripting": "1.5.2",
+    "com.unity.visualscripting": "1.7.6",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
