+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.1.6f1
-m_EditorVersionWithRevision: 2021.1.6f1 (c0fade0cc7e9)
+m_EditorVersion: 2021.2.11f1
+m_EditorVersionWithRevision: 2021.2.11f1 (e50cafbb4399)
